rename Card.jsx to CardGallery.jsx
